refactor: update tests with new import links
