+++ product-service/test/getProductById.test.ts
@@ -3,9 +3,9 @@ import {
   APIGatewayProxyResult,
   Context,
 } from "aws-lambda";
-import { handler } from "../lambda-functions/getProductById";
+import { handler } from "../lambda-functions/handlers/getProductById";
 import { products } from "../lambda-functions/products";
-import { IProduct } from "../lambda-functions/product.interface";
+import { IProduct } from "../lambda-functions/handlers/product.interface";
 
 describe("getProductById lambda function", (): void => {
   beforeEach((): void => {
